requirements.txt
Django==4.0.4
djangorestframework==3.13.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
Pillow==9.0.1
Jinja2==3.1.4
whitenoise==6.7.0
django-filter==22.1